+++ server/wrapper/nmea/de.uniol.inf.is.odysseus.wrapper.nmea/src/de/uniol/inf/is/odysseus/wrapper/nmea/sentence/aissentences/payload/decoded/PositionReportClassA.java
@@ -23,8 +23,8 @@ public class PositionReportClassA extends PositionReport {
 	public PositionReportClassA(String originalNmea, AISMessageType messageType,
 			Integer repeatIndicator, MMSI sourceMmsi,
 			NavigationStatus navigationStatus, Integer rateOfTurn,
-			Float speedOverGround, Boolean positionAccurate, Double latitude,
-			Double longitude, Float courseOverGround, Integer trueHeading,
+			Double speedOverGround, Boolean positionAccurate, Double latitude,
+			Double longitude, Double courseOverGround, Integer trueHeading,
 			Integer second, ManeuverIndicator maneuverIndicator,
 			Boolean raimFlag) {
 		super(originalNmea, messageType, repeatIndicator, sourceMmsi, navigationStatus, rateOfTurn,
